Fundamental Kotlin - Overriding.
